+++ .ps1
@@ -1,0 +1,2 @@
+﻿# debug profile
+Import-Module Ldbc